requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pg8000==1.30.3
python-dateutil==2.8.2
openai==1.3.0
